# Conflicts:
#	src/main/java/ru/skypro/homework/WebSecurityConfig.java
#	src/main/java/ru/skypro/homework/controller/AdsController.java
#	src/main/java/ru/skypro/homework/dto/UserSecurity.java
#	src/main/java/ru/skypro/homework/mapper/UserMapper.java
#	src/main/java/ru/skypro/homework/security/MyUserDetails.java
#	src/main/java/ru/skypro/homework/security/MyUserDetailsService.java
#	src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
#	src/main/java/ru/skypro/homework/service/impl/AuthServiceImpl.java

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -11,6 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.Ads;
 import ru.skypro.homework.dto.CreateAds;
 import ru.skypro.homework.dto.FullAds;
+import ru.skypro.homework.dto.Role;
 import ru.skypro.homework.dto.ResponseWrapperAds;
 import ru.skypro.homework.entity.AdsEntity;
 import ru.skypro.homework.entity.ImageEntity;
@@ -20,14 +21,13 @@ import ru.skypro.homework.repository.AdsRepository;
 import ru.skypro.homework.repository.UserRepository;
 import ru.skypro.homework.security.MyUserDetails;
 import ru.skypro.homework.service.AdsService;
-import ru.skypro.homework.service.CommentService;
 import ru.skypro.homework.service.ImageService;
 import ru.skypro.homework.service.UserService;
+import ru.skypro.homework.service.CommentService;
 
 import java.io.IOException;
 import java.util.Collection;
-
-import static org.springframework.util.ObjectUtils.isEmpty;
+import java.util.Optional;
 
 /**
  * Сервис AdsServiceImpl
@@ -62,7 +62,9 @@ public class AdsServiceImpl implements AdsService {
     private final UserService userService;
     private final ImageService imageService;
     private final CommentService commentService;
-    private final MyUserDetails myUserDetails; // string security положит сюда авторизированного пользователя
+
+    private final MyUserDetails userDetails;
+
 
     /**
      * Конструктор - создание нового объекта репозитория
@@ -71,6 +73,7 @@ public class AdsServiceImpl implements AdsService {
      * @param adsMapper
      * @param userRepository репозиторий пользователя
      * @param userService    сервис пользователя
+     * @param userMapper
      * @param imageService
      * @param commentService
      * @param myUserDetails
@@ -96,13 +99,11 @@ public class AdsServiceImpl implements AdsService {
     @Override
     public Collection<Ads> getAllAds(String title) {
         logger.info("Вызван метод получения всех объявлений");
-        Collection<AdsEntity> adsEntities;
-        if (!isEmpty(title)) {
-            adsEntities = adsRepository.findByTitleLikeIgnoreCase(title);
-        } else {
-            adsEntities = adsRepository.findAll();
+        if (title == null) {
+
+            return adsMapper.adsEntityToCollectionDto(adsRepository.findAll());
         }
-        return adsMapper.adsEntityToCollectionDto(adsEntities);
+        return adsMapper.adsEntityToCollectionDto(adsRepository.findByTitleLikeIgnoreCase(title));
     }
 
     /**
@@ -110,18 +111,19 @@ public class AdsServiceImpl implements AdsService {
      *
      * @param createAds      данные объявления
      * @param image          картинка объявления
-
+     * @param authentication авторизованный пользователь
      * @return добавленное новое объявление
      */
     @Override
-    public Ads createAds(CreateAds createAds, MultipartFile image) {
+    public Ads createAds(CreateAds createAds, MultipartFile image, Authentication authentication) {
         if (createAds.getPrice() < 0) {
             throw new IllegalArgumentException("Цена должна быть больше 0!");
         }
         logger.info("Вызван метод добавления объявления");
 
         AdsEntity adsEntity = adsMapper.toEntity(createAds);
-        adsEntity.setAuthor(userService.getNameUser(myUserDetails.getUsername()));
+        UserEntity author = userRepository.findByEmailIgnoreCase(authentication.getName()).orElseThrow(RuntimeException::new); //TODO сделать свое исключение
+        adsEntity.setAuthor(author);
 
         ImageEntity adImage;
         try {
@@ -140,39 +142,43 @@ public class AdsServiceImpl implements AdsService {
     /**
      * Получение объявления по идентификатору (id), хранящихся в базе данных
      *
-     * @param adId идентификатор объявления, не может быть null
+     * @param adsId идентификатор объявления, не может быть null
      * @return возвращает объявление по идентификатору (id)
      */
     @Override
-    public FullAds getAds(Integer adId) {
+    public FullAds getAds(Integer adsId) {
         logger.info("Вызван метод получения объявления по идентификатору (id)");
-        return adsMapper.toFullAdsDto(adsRepository.findById(adId).orElseThrow()); // TODO сделать свое исключение
+        return adsMapper.toFullAdsDto(adsRepository.findById(adsId).orElseThrow());
     }
 
     /**
      * Удаление объявления по идентификатору (id), хранящихся в базе данных
      *
-     * @param adId идентификатор объявления, не может быть null
+     * @param adsId          идентификатор объявления, не может быть null
      */
     @Override
-    public void deleteAds(Integer adId) {
+    public void deleteAds(Integer adsId) {
         logger.info("Вызван метод удаления объявления по идентификатору (id)");
-        AdsEntity adsEntity = adsRepository.findById(adId).orElseThrow(); // TODO сделать свое исключение
-        commentService.deleteAllByAdsId(adId);
-        adsRepository.deleteById(adId);
-        imageService.deleteImage(adsEntity.getImageEntity().getId());
+        AdsEntity deleteAd = adsRepository.findById(adsId).orElseThrow(RuntimeException::new);
+        if (deleteAd.getAuthor().getEmail().equals(userDetails.getUserSecurity().getEmail()) || userDetails.getUserSecurity().getRole() == Role.ADMIN) {
+            commentService.deleteAllByAdsId(adsId);
+            adsRepository.deleteById(adsId);
+            imageService.deleteImage(adsEntity.getImageEntity().getId());
+        } else {
+            throw new RuntimeException("Вы не можете удалять чужие объявления");
+        }
     }
 
     /**
      * Обновление объявления по идентификатору (id), хранящихся в базе данных
      *
-     * @param adId     идентификатор объявления, не может быть null
-     * @param createAds данные объявления
+     * @param adsId          идентификатор объявления, не может быть null
+     * @param createAds      данные объявления
      * @return возвращает обновленное объявление по идентификатору (id)
      */
     @Override
-    public Ads updateAds(CreateAds createAds, Integer adId) {
-        if (adId == null) {
+    public Ads updateAds(CreateAds createAds, Integer adsId) {
+        if (adsId == null) {
             throw new RuntimeException("Такого объявления не существует!");
         }
 
@@ -180,40 +186,43 @@ public class AdsServiceImpl implements AdsService {
             throw new RuntimeException("Цена должна быть больше 0!");
         }
 
-        AdsEntity updateAd = adsRepository.findById(adId).orElseThrow(RuntimeException::new);
-        updateAd.setTitle(createAds.getTitle());
-        updateAd.setPrice(createAds.getPrice());
-        updateAd.setDescription(createAds.getDescription());
+        AdsEntity updateAd = adsRepository.findById(adsId).orElseThrow(RuntimeException::new);
+        if (updateAd.getAuthor().getEmail().equals(userDetails.getUserSecurity().getEmail()) || userDetails.getUserSecurity().getRole() == Role.ADMIN) {
+            updateAd.setTitle(createAds.getTitle());
+            updateAd.setPrice(createAds.getPrice());
+            updateAd.setDescription(createAds.getDescription());
 
-        adsRepository.save(updateAd);
+            adsRepository.save(updateAd);
 
-        return adsMapper.toAdsDto(updateAd);
+            return adsMapper.toAdsDto(updateAd);
+        }
+        throw new RuntimeException("Вы не можете изменять чужие объявления");
     }
 
     /**
      * Получение объявлений авторизованного пользователя, хранящихся в базе данных
      *
+     * @param authentication авторизованный пользователь
      * @return возвращает все объявления авторизованного пользователя
      */
     @Override
-    public Collection<Ads> getAdsMe() {
+    public Collection<Ads> getAdsMe(Authentication authentication) {
         logger.info("Вызван метод получения объявлений авторизованного пользователя");
-        UserEntity userEntity = userService.getNameUser(myUserDetails.getUsername());
-        Collection<AdsEntity> adsEntity = adsRepository.findAllAdsByAuthorId(userEntity.getId());
-        return adsMapper.adsEntityToCollectionDto(adsEntity);
+        Collection<AdsEntity> AdsEntity = adsRepository.findAllAdsByAuthorId(userService.getUser(authentication).getId());
+        return adsMapper.adsEntityToCollectionDto(AdsEntity);
     }
 
     /**
      * Обновление картинки объявления
      *
-     * @param adId  идентификатор объявления, не может быть null
-     * @param image картинка объявления
+     * @param adsId          идентификатор объявления, не может быть null
+     * @param image          картинка объявления
      * @return объявление с новой картинкой
      */
     @Override
-    public String updateImage(Integer adId, MultipartFile image) {
+    public String updateImage(Integer adsId, MultipartFile image) {
         logger.info("Вызван метод обновления картинки объявления");
-        if (adId == null) {
+        if (adsId == null) {
             throw new RuntimeException("Такого объявления не существует!");
         }
 
@@ -224,13 +233,18 @@ public class AdsServiceImpl implements AdsService {
             throw new RuntimeException("Ошибка при загрузке фото");
         }
 
-        int imageId = adImage.getId();
-        imageService.deleteImage(imageId);
+        AdsEntity updateAd = adsRepository.findById(adsId).orElseThrow(RuntimeException::new);
+        if (updateAd.getAuthor().getEmail().equals(userDetails.getUserSecurity().getEmail()) || userDetails.getUserSecurity().getRole() == Role.ADMIN) {
 
-        AdsEntity ad = adsRepository.findById(adId).orElseThrow(RuntimeException::new);
-        ad.setImageEntity(adImage);
-        adsRepository.save(ad);
-        return adsMapper.toAdsDto(ad).getImage();
+            int imageId = adImage.getId();
+            imageService.deleteImage(imageId);
+
+            updateAd.setImageEntity(adImage);
+            adsRepository.save(updateAd);
+            return adsMapper.toAdsDto(updateAd).getImage();
+        }
+
+        throw new RuntimeException("Вы не можете изменять чужие объявления");
     }
 
     /**
